10.3 fixed the percentages

+++ V2X-Java/src/main/java/v2x/AnswerCounter.java
@@ -33,9 +33,9 @@ public class AnswerCounter {
      *
      * @return <code>int[]</code> an array of the percentages
      */
-    public int[] getPercentage() {
-        int[] answer = new int[3];
-        int totalAnswers = this.answerZero + this.answerOne + this.answerTwo;
+    public double[] getPercentage() {
+        double[] answer = new double[3];
+        double totalAnswers = this.answerZero + this.answerOne + this.answerTwo;
         answer[0] = (this.answerZero / totalAnswers) * 100;
         answer[1] = (this.answerOne / totalAnswers) * 100;
         answer[2] = (this.answerTwo / totalAnswers) * 100;
@@ -47,7 +47,7 @@ public class AnswerCounter {
      * Prints the math related answers
      */
     public void printMath() {
-        int[] answer = getPercentage();
+        double[] answer = getPercentage();
         for (int i = 0; i < answer.length; i++) {
             System.out.println("Percentage of answer type " + (i + 1) + ":");
             System.out.println(answer[i]);
